requirements.txt
# requirements.txt
# install with: pip install -r requirements.txt

discord.py>=2.3.0     # the discord library — needs 2.x for timeout support
aiosqlite>=0.19.0     # async SQLite so database calls don't block the bot
python-dotenv>=1.0.0  # loads my .env file for secrets
